+++ acc_app_optimisation/gui/configuration/_dialog.py
@@ -2,7 +2,6 @@
 
 import logging
 import typing as t
-from types import SimpleNamespace
 
 import gym
 import numpy as np
@@ -10,9 +9,15 @@ from cernml import coi
 from PyQt5.QtCore import pyqtSignal
 from PyQt5.QtWidgets import QDialog, QDialogButtonBox, QTabWidget, QVBoxLayout, QWidget
 
+from ...utils.disabled_updates import disabled_updates
+from ...utils.coerce_float import coerce_float_tuple
 from ...utils.typecheck import AnyOptimizable, is_configurable, is_function_optimizable
 from ..excdialog import exception_dialog
-from ._skeleton_points import SkeletonPointsWidget
+from ._skeleton_points import (
+    BaseSkeletonPointsWidget,
+    SkeletonPointsEditWidget,
+    SkeletonPointsViewWidget,
+)
 from ._widget import ConfigureWidget
 
 LOG = logging.getLogger(__name__)
@@ -22,8 +27,7 @@ class _BaseDialog(QDialog):
     """Common logic of `PureConfigureDialog` and `ProblemConfigureDialog`.
 
     Args:
-        target: The environment to be configured. If None is passed, no
-            `ConfigureWidget` is created.
+        target: The environment to be configured.
         parent: The parent widget to attach to.
 
     Attributes:
@@ -33,12 +37,13 @@ class _BaseDialog(QDialog):
 
     config_applied = pyqtSignal()
 
-    def __init__(
-        self, target: t.Optional[coi.Configurable], parent: t.Optional[QWidget] = None
-    ) -> None:
+    def __init__(self, target: t.Any, parent: t.Optional[QWidget] = None) -> None:
         super().__init__(parent)
+        self.setWindowTitle(f"Configuring {_get_configurable_name(target)} …")
         self.target = target
-        self._cfgform = None if target is None else ConfigureWidget(target.get_config())
+        self._cfgform = (
+            ConfigureWidget(target.get_config()) if is_configurable(target) else None
+        )
         self._controls = QDialogButtonBox(  # type: ignore
             QDialogButtonBox.Ok | QDialogButtonBox.Apply | QDialogButtonBox.Cancel
         )
@@ -47,10 +52,6 @@ class _BaseDialog(QDialog):
             self._on_apply_clicked
         )
         self._controls.button(QDialogButtonBox.Cancel).clicked.connect(self.reject)
-        if target is not None:
-            self.setWindowTitle(f"Configuring {_get_configurable_name(target)} …")
-        else:
-            self.setWindowTitle("Configuring …")
 
     def _on_ok_clicked(self) -> None:
         """Apply the configs and close the window."""
@@ -71,9 +72,9 @@ class _BaseDialog(QDialog):
             True if the config has been applied successfully. False if
             an exception has been raised.
         """
-        if self.target is None:
+        if self._cfgform is None:
             return True
-        assert self._cfgform is not None
+        assert is_configurable(self.target), self.target
         try:
             values = self._cfgform.current_values()
             self.target.apply_config(values)
@@ -111,42 +112,39 @@ class OptimizableDialog(_BaseDialog):
         parent: The parent widget to attach to.
     """
 
+    # pylint: disable = invalid-name
+
     def __init__(
         self,
         target: AnyOptimizable,
-        skeleton_points: t.Optional[np.ndarray] = None,
+        skeleton_points: t.Tuple[float, ...] = (),
         parent: t.Optional[QWidget] = None,
     ) -> None:
-        if is_configurable(target):
-            super().__init__(target, parent)
-        else:
-            super().__init__(None, parent)
-            self.setWindowTitle(f"Configuring {_get_configurable_name(target)} …")
-        self._points_page: t.Optional[SkeletonPointsWidget]
-        tab_widget = QTabWidget()
+        super().__init__(target, parent)
+        self._points_page: t.Optional[BaseSkeletonPointsWidget]
+        self._tab_widget = QTabWidget()
         if self._cfgform is not None:
-            tab_widget.addTab(self._cfgform, "Configuration")
-        self._skeleton_points: t.Optional[np.ndarray]
+            self._tab_widget.addTab(self._cfgform, "Configuration")
         if is_function_optimizable(target):
-            if skeleton_points is not None:
-                self._skeleton_points = np.array(skeleton_points, dtype=float)
+            points_override = target.override_skeleton_points()
+            if points_override is not None:
+                self._skeleton_points = coerce_float_tuple(points_override)
+                self._points_page = SkeletonPointsViewWidget(self._skeleton_points)
             else:
-                self._skeleton_points = np.array([], dtype=float)
-            self._points_page = SkeletonPointsWidget(self._skeleton_points)
-            tab_widget.addTab(self._points_page, "Skeleton points")
+                self._skeleton_points = skeleton_points
+                self._points_page = SkeletonPointsEditWidget(self._skeleton_points)
+            self._tab_widget.addTab(self._points_page, "Skeleton points")
         else:
             self._points_page = None
-            self._skeleton_points = None
+            self._skeleton_points = ()
         main_layout = QVBoxLayout(self)
-        main_layout.addWidget(tab_widget)
+        main_layout.addWidget(self._tab_widget)
         main_layout.addWidget(self._controls)
 
-    def skeletonPoints(self) -> t.Optional[np.ndarray]:  # pylint: disable=invalid-name
+    def skeletonPoints(self) -> t.Tuple[float, ...]:
         return self._skeleton_points
 
-    def setSkeletonPoints(  # pylint: disable=invalid-name
-        self, points: np.ndarray
-    ) -> None:
+    def setSkeletonPoints(self, points: t.Tuple[float, ...]) -> None:
         if self._points_page is not None:
             self._skeleton_points = points
             self._points_page.setSkeletonPoints(points)
@@ -156,13 +154,49 @@ class OptimizableDialog(_BaseDialog):
             )
 
     def apply_config(self) -> bool:
-        if self._points_page is not None:
+        if isinstance(self._points_page, SkeletonPointsViewWidget):
+            assert is_function_optimizable(self.target), self.target
+            # Skeleton points page is read-only -- update it.
+            # But only if the configs were updated successfully.
+            success = super().apply_config()
+            points_override = self.target.override_skeleton_points()
+            if points_override is None:
+                self._replace_skeleton_points_widget(
+                    SkeletonPointsEditWidget(self._skeleton_points)
+                )
+            else:
+                self._points_page.setSkeletonPoints(coerce_float_tuple(points_override))
+            return success
+        if isinstance(self._points_page, SkeletonPointsEditWidget):
+            assert is_function_optimizable(self.target), self.target
+            # Skeleton points page is editable -- read its values.
+            # If that succeeds, continue with regular configs.
             try:
                 self._skeleton_points = self._points_page.skeletonPoints()
             except ValueError as exc:
                 _show_skeleton_points_failed(exc, parent=self)
                 return False
+            success = super().apply_config()
+            points_override = self.target.override_skeleton_points()
+            if points_override is not None:
+                self._skeleton_points = coerce_float_tuple(points_override)
+                self._replace_skeleton_points_widget(
+                    SkeletonPointsViewWidget(self._skeleton_points)
+                )
+            return success
+        # There is no skeleton points page. Just delegate to configs.
+        assert self._points_page is None
         return super().apply_config()
+
+    def _replace_skeleton_points_widget(self, new: BaseSkeletonPointsWidget) -> None:
+        assert self._points_page is not None
+        with disabled_updates(self._tab_widget) as tabs:
+            current_index = tabs.currentIndex()
+            points_tab_index = tabs.indexOf(self._points_page)
+            tabs.removeTab(points_tab_index)
+            new_index = tabs.insertTab(points_tab_index, new, "Skeleton points")
+            if current_index == points_tab_index:
+                tabs.setCurrentIndex(new_index)
 
 
 class EnvDialog(_BaseDialog):
@@ -209,7 +243,7 @@ class ConfigTimeLimit(gym.Wrapper, coi.Configurable):
         )
         return config
 
-    def apply_config(self, values: SimpleNamespace) -> None:
+    def apply_config(self, values: coi.ConfigValues) -> None:
         self.value = values.TimeLimit_max_episode_steps
         if is_configurable(self.env):
             self.env.apply_config(values)