+++ acc_app_optimisation/job_control/single_objective/__init__.py
@@ -4,6 +4,7 @@ from . import optimizers
 from .builder import CannotBuildJob, OptJobBuilder
 from .jobs import OptJob
 from .optimizers import ALL_OPTIMIZERS, OptimizerFactory
+from .skeleton_points import SkeletonPoints, gather_skeleton_points
 
 __all__ = [
     "ALL_OPTIMIZERS",
@@ -11,5 +12,7 @@ __all__ = [
     "OptJob",
     "OptJobBuilder",
     "OptimizerFactory",
+    "SkeletonPoints",
+    "gather_skeleton_points",
     "optimizers",
 ]